+++ clang-tools-extra/test/clang-tidy/checkers/misc-alphabetical-decl-order.cpp
@@ -1,0 +1,16 @@
+// RUN: %check_clang_tidy %s misc-alphabetical-decl-order %t
+
+// FIXME: Add something that triggers the check here.
+typedef struct node {
+    int data;
+    node* next;
+} node;
+
+void f() {
+    //Not ok
+    int b, a, c;
+    node bye, hi;
+
+    //Ok
+    int a, b, c;
+}